+++ packages/terra-navigation-side-menu/tests/wdio/navigation-side-menu-spec.js
@@ -10,6 +10,10 @@ describe('Default navigation side menu display', () => {
   Terra.should.matchScreenshot('#test-menu', { selector: '#test-menu' });
   Terra.should.beAccessible({ context: '#test-menu' });
   Terra.should.themeEachCustomProperty('#test-menu', {
+    '--terra-navigation-side-menu-list-item-border-bottom': '5px solid #f00',
+    '--terra-navigation-side-menu-list-item-border-left': '5px solid #f00',
+    '--terra-navigation-side-menu-list-item-border-right': '5px solid #f00',
+    '--terra-navigation-side-menu-list-item-border-top': '5px solid #f00',
     '--terra-navigation-side-menu-item-color': 'green',
     '--terra-navigation-side-menu-item-size': '0.5rem',
     '--terra-navigation-side-menu-item-padding-top': '0',
@@ -30,6 +34,7 @@ describe('Hover navigation side menu display', () => {
   Terra.should.matchScreenshot('#test-menu', { selector: '#test-menu' });
   Terra.should.themeEachCustomProperty('#test-menu', {
     '--terra-navigation-side-menu-item-hover-background-color': 'orange',
+    '--terra-navigation-side-menu-item-hover-color': 'blue',
   });
 });
 
@@ -47,6 +52,7 @@ describe('Selected navigation side menu display', () => {
   Terra.should.beAccessible({ context: '#test-menu' });
   Terra.should.themeEachCustomProperty('#test-menu', {
     '--terra-navigation-side-menu-item-selected-background-color': 'pink',
+    '--terra-navigation-side-menu-item-selected-color': 'black',
     '--terra-navigation-side-menu-item-selected-border-left': '1rem dotted orange',
     '--terra-navigation-side-menu-item-selected-padding-left': '0',
   });